# [perf] Fast color manipulation