@romgrk 2024-08-17
I spent some more time refining the approach and optimizing the parsing/formatting functions. There was an issue with JS bitwise operators as they operate on 32-bits signed, and V8's Smi (small integer) number representation is also a 32-bits signed value, so it's just much faster to use 32-bits signed. As far as bitwise manipulation is concerned it's not an issue as the bits are the same for signed/unsigned integers. The only disadvantage is that our representation now contains negative numbers, e.g. `#ffffffff` is represented by the number `-1` instead of the number `0xffffffff` (even though both are stored as the bits `0b11111...1111` due to two's complement). It's not that much an issue for us here as our manipulation functions are a `string -> string` transformation, so the underlying representation is hidden from the caller at all times.

Anyway, with all those improvements here are the new results. I have [updated the test case](https://gist.github.com/romgrk/ade3ee62b4641d5b375127e98d354b5d) a bit and in this one there are less buttons and more of other elements. That change in itself dropped the color manipulation runtime to around 7.5% of the total in the **before** case (which only means that buttons do more color manipulation than the rest of the components).  And for the **after** case, the new approach is so fast that no color manipulation function even appears in the stack traces (verified a few times), which means  it's now lower than 0.01% of the total runtime.

| before (7.8%) | after (<0.01%) |
| --- | --- |
| ![color-before](https://github.com/user-attachments/assets/676c0a33-a9c3-46bc-9d49-9173166b5221) | ![color-after](https://github.com/user-attachments/assets/18d75410-e14d-46fa-96d5-60a2ded099c9) |

---

### Comparing other packages

I also benchmarked this approach against other existing packages on NPM using `colord`'s test suite. Unsurprisingly it's very fast, being nearly 8x faster than the 2nd fastest library (`colord`).

```
Running "Parse hexadecimal, set to 0.5 opacity, and convert back to hexadecimal" suite...

  colord:
    26 992 ops/s, ±0.73%    | 85.14% slower
  color:
    6 075 ops/s, ±0.80%     | 96.66% slower
  tinycolor2:
    9 907 ops/s, ±0.77%     | 94.55% slower
  chroma-js:
    7 036 ops/s, ±0.82%     | 96.13% slower
  MUI colorManipulator:
    5 111 ops/s, ±0.93%     | slowest, 97.19% slower
  MUI color bitwise:
    181 652 ops/s, ±0.75%   | fastest

Finished 6 cases!
  Fastest: MUI color bitwise
  Slowest: MUI colorManipulator
```

The test case is however a bit contrived, as it goes through hexadecimal parsing and formatting which have been optimized *a lot*. For `rgb(50, 100, 200)` color formats, this approach is *only* around 2x faster than the 2nd fastest (also `colord`).