@oliviertassinari 2024-08-18
> Those functions run everytime the component mounts (and renders I think), but the implementation is very costly

The constraint and the dimensions we should optimize for are interesting here. It has been a moving target.

**The past (<= v4)**
Before we depended on Emotion, we depended on JSS. The code would call those color transformations a handful of time. Effectively once per component, not once per render. There, our main constraint was bundle size. I believe runtime cost wasn't much of a concern.

**The present (v5)**
Since we released Material UI v5, we have been aware that our implementation on top of Emotion is slow (emotion itself not being fast but I'm talking only about our wrapper), and I have been encouraging us to solve those performance issues. The system is slow, our wrapper on top of Emotion's styled is slow etc. We know this because when we benchmark with equivalent wrapper, we didn't saw the same slowness. It's in the order or 10-30% slower.

So unlike in v4, we need to care about runtime and bundle size, and can sacrifice a bit more bundle size.

At the same time, the CSS specification evolved. They introduced: 1. color space, 2. new syntax for the colors, and 3. CSS variables. I'm not aware we did much work to support those. We need to.

**The future**

We are moving toward removing those color transformations JavaScript helpers: https://github.com/mui/material-ui/issues/43180. This is relevant for our runtime and non-runtime version of our style library. I think that the sooner we can do it, the better.

I have no idea if static extraction will be successful:

- If it fails, and we move to CSS Modules, I imaging that we would apply CSS Variables for the theme values. If we move to Tailwind CSS, same. So this seems invariant. 
- If it's successful. The pressure for performance is lower than in the runtime but still present, developers build needs to be fast. The presence for smaller bundle size is almost non existant, just the npm package size.
- If it's successful and we have both a runtime and non runtime (the most likely outcome IMHO). Then we should factor both constraints into account.