@oliviertassinari 2024-08-18
> I would like to propose to replace the color manipulation functions with this new approach.

Improving the performance of the test case by 8% just by working on the color transformations helper is great to see! I love to see focus on optimizing the runtime of the core.

(A side note, when I see https://twind.dev/handbook/introduction.html#benchmarks and goober so fast, maybe we don't even need static css extraction, maybe can both solve RSC, streaming, and performance without it.)

Question, from my end: What's the bundle size cost?