@romgrk 2024-08-18
For the bundle size, it looks to be around 4KB. Feature-wise this PR is nearly complete so I'm not going to add any substantial amount of code, in fact I still need to remove the old color manipulation code, so it should stay at 4KB at most, possibly less.

```
108K   ├── bundle-current.min.js.gz
112K   ├── bundle-new.min.js.gz
```

The biggest cost is parsing non-sRGB color spaces, if we only supported hex, rgb, hsl, hwb, we'd probably be around 2KB.