@siriwatknp 2024-08-19
> only supported hex, rgb, hsl, hwb

I think we should keep it like that. This should be a perf improvement, not a new feature to add more support. In the future, it's likely that we will fade away from JS according to my [POC](https://github.com/mui/material-ui/pull/41237)