@mnajdova 2024-08-19
Thanks for looking into this Romain. The improvements is too big for us to not look into it. As mentioned before, we are likely going to have Emotion for a while, so this improvements are going to make a difference. 👌 

If we can't finish this on time for v6, we can always release it later, by deprecating the old color manipulators and exposing the new ones (unless the plan is to keep the API identical).