@siriwatknp 2024-08-19
> If we can't finish this on time for v6, we can always release it later, by deprecating the old color manipulators and exposing the new ones (unless the plan is to keep the API identical).

For the release, I think this can be a minor of v6. Let's not include this before the stable to give some time to discuss and test.

In my opinion, the first step should be identical API to improve performance.