@romgrk 2024-08-19
If the objective is compatibility with what `colorManipulator` currently supports (and not compatibility with the CSS spec), then this PR already does more than that and I can trim some of the code.

I've also tested it by swapping the internals of `colorManipulator` [like this](https://github.com/mui/material-ui/pull/43289/files#diff-edf2b8504e7feb81430c325c166d5484dc11adbff80eef5ffbac9ffcae65a002R244-R247), and as far as I can see there isn't any breaking API change needed. The test file `colorManipulator.test.js` can remain as it is, and the new `color.test.js` also adds a few more tests of its own.